+++ kart/repo.py
@@ -293,13 +293,13 @@ class KartRepo(pygit2.Repository):
                 "-c",
                 "remote.origin.fetch=+refs/filters/*:refs/filters/*",
             ]
-            extension_spec = spatial_filter_spec.git_spatial_filter_extension_spec()
-            if extension_spec and os.environ.get("X_KART_SPATIAL_FILTERED_CLONE"):
+            partial_clone_spec = spatial_filter_spec.partial_clone_filter_spec()
+            if partial_clone_spec and os.environ.get("X_KART_SPATIAL_FILTERED_CLONE"):
                 click.echo(
-                    f"Cloning using git spatial filter extension: {extension_spec}",
+                    f"Cloning using git spatial filter extension: {partial_clone_spec}",
                     err=True,
                 )
-                extra_args.append(extension_spec)
+                extra_args.append(f"--filter={partial_clone_spec}")
 
         if bare:
             kart_repo = cls._create_with_git_command(